% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_all_bind_variables}
\alias{get_all_bind_variables}
\title{Get the bind variables in sql statement}
\usage{
get_all_bind_variables(sql)
}
\arguments{
\item{sql}{statement}
}
\value{
all bind variables in the sql statement
}
\description{
This function takes sql statement and returns all the bind variables in the sql statement.
}
\examples{
get_all_bind_variables("select * from users where userid = :bind_userid")
}
