% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_all_select_cols_with_alias}
\alias{get_all_select_cols_with_alias}
\title{Get the names of the selected columns in the sql statement}
\usage{
get_all_select_cols_with_alias(sql)
}
\arguments{
\item{sql}{statement}
}
\value{
names of columns selected in sql statement
}
\description{
This function takes sql statement and returns all the columns selected in the sql
statement.
}
\examples{
get_all_select_cols_with_alias("Select p from abc where xyz = 5")
}
