% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_all_subqueries}
\alias{get_all_subqueries}
\title{Get the subqueries in sql statement}
\usage{
get_all_subqueries(sql)
}
\arguments{
\item{sql}{statement}
}
\value{
all subqueries for the sql statement
}
\description{
This function takes sql statement and returns the list of subqueries.It does not
return nested sub queries. For nested subqueries, each subquery needs to be fed
into the method again.
}
\examples{
get_all_subqueries("select * from users where appid in
  (select appid from applications)")
}
