% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_all_tables_with_alias}
\alias{get_all_tables_with_alias}
\title{Get the names of the tables with alias present in the sql}
\usage{
get_all_tables_with_alias(sql)
}
\arguments{
\item{sql}{statement}
}
\value{
names of tables with alias present in sql statement
}
\description{
This function takes sql statement and returns the list of names
of tables and alias for that table, if any present in the sql statement.
The first column represent table name and second column represents alias.
}
\examples{
get_all_tables_with_alias("Select p from abc where xyz = 5")
}
